chore: refactor event type imports and rename event-types to solo-event-type
