+++ src/commands/one-shot/default-one-shot.ts
@@ -84,7 +84,7 @@ import {MirrorNodeStateSchema} from '../../data/schema/model/remote/state/mirror
 import {ExplorerStateSchema} from '../../data/schema/model/remote/state/explorer-state-schema.js';
 import {BlockNodeStateSchema} from '../../data/schema/model/remote/state/block-node-state-schema.js';
 import {type SoloEventBus} from '../../core/events/solo-event-bus.js';
-import {SoloEventType} from '../../core/events/event-types/event-types.js';
+import {SoloEventType} from '../../core/events/event-types/solo-event.js';
 import {MirrorNodeDeployedEvent} from '../../core/events/event-types/mirror-node-deployed-event.js';
 import {NodesStartedEvent} from '../../core/events/event-types/nodes-started-event.js';
 import {DeploymentSchema} from '../../data/schema/model/local/deployment-schema.js';
