+++ src/core/events/event-types/network-deployed-event.ts
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-import {SoloEventType} from './event-types.js';
-import {SoloEvent} from './solo-event.js';
+import {SoloEvent, SoloEventType} from './solo-event.js';
 
 export class NetworkDeployedEvent extends SoloEvent {
   public constructor(public readonly deployment: string) {
