+++ src/core/events/event-types/solo-event-type.ts
@@ -4,10 +4,4 @@ import {type NetworkDeployedEvent} from './network-deployed-event.js';
 import {type MirrorNodeDeployedEvent} from './mirror-node-deployed-event.js';
 import {type NodesStartedEvent} from './nodes-started-event.js';
 
-export enum SoloEventType {
-  NetworkDeployed = 'NetworkDeployed',
-  NodesStarted = 'NodesStarted',
-  MirrorNodeDeployed = 'MirrorNodeDeployed',
-}
-
 export type AnySoloEvent = NodesStartedEvent | NetworkDeployedEvent | MirrorNodeDeployedEvent;
