+++ src/core/events/event-types/solo-event.ts
@@ -1,7 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
-import {type SoloEventType} from './event-types.js';
+export enum SoloEventType {
+  NetworkDeployed = 'NetworkDeployed',
+  NodesStarted = 'NodesStarted',
+  MirrorNodeDeployed = 'MirrorNodeDeployed',
+}
 
 export abstract class SoloEvent {
-  public constructor(public readonly type: SoloEventType) {}
+  protected constructor(public readonly type: SoloEventType) {}
 }
